chore(back): updates to and removes patches for strapi@14.5.5 (#1033)
Closes: #1033
Fixes: https://github.com/strapi/strapi/issues/18665
