+++ back/package.json
@@ -16,14 +16,18 @@
     "strapi": "strapi"
   },
   "dependencies": {
-    "@strapi/plugin-i18n": "4.14.6",
-    "@strapi/plugin-sentry": "4.14.6",
+    "@strapi/plugin-i18n": "4.15.5",
+    "@strapi/plugin-sentry": "4.15.5",
     "@strapi/plugin-seo": "^1.9.4",
-    "@strapi/plugin-users-permissions": "4.14.6",
-    "@strapi/provider-upload-aws-s3": "4.14.6",
-    "@strapi/strapi": "4.14.6",
+    "@strapi/plugin-users-permissions": "4.15.5",
+    "@strapi/provider-upload-aws-s3": "4.15.5",
+    "@strapi/strapi": "4.15.5",
     "pg": "^8.11.3",
-    "strapi-plugin-config-sync": "^1.2.3"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^5.3.4",
+    "strapi-plugin-config-sync": "^1.2.3",
+    "styled-components": "^5.3.11"
   },
   "devDependencies": {
     "patch-package": "8.0.0"
