--- back/patches/@strapi+strapi+4.14.6.patch
@@ -1,23 +1,0 @@
-diff --git a/node_modules/@strapi/strapi/dist/services/webhook-runner.js b/node_modules/@strapi/strapi/dist/services/webhook-runner.js
-index 591b250..bd9d014 100644
---- a/node_modules/@strapi/strapi/dist/services/webhook-runner.js
-+++ b/node_modules/@strapi/strapi/dist/services/webhook-runner.js
-@@ -12,6 +12,7 @@ const worker_queue_1 = __importDefault(require("./worker-queue"));
- const debug = (0, debug_1.default)('strapi:webhook');
- const defaultConfiguration = {
-     defaultHeaders: {},
-+    extraBody: {},
- };
- class WebhookRunner {
-     eventHub;
-@@ -68,9 +69,7 @@ class WebhookRunner {
-         return this.fetch(url, {
-             method: 'post',
-             body: JSON.stringify({
--                event,
--                createdAt: new Date(),
--                ...info,
-+                ...this.config.extraBody,
-             }),
-             headers: {
-                 ...this.config.defaultHeaders,
